Change root with up-dir shortcuts. Fixes jaypei#290
Currently the shortcut keys (RET, SPC) do not trigger a change root to the parent directory on the ".." node. To trigger it, you have to move your cursor to the ".." button.

This fix makes the shortcut keys trigger a change root if the path is equal to the parent directory path.